pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "API Django com DRF"
authors = ["Francisco Lucas <franciscolucas686@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.2.4"
djangorestframework = "^3.16.0"
django-extensions = "^4.1"
django-debug-toolbar = "^5.2.0"
psycopg2 = "^2.9"

[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
